+++ app/components/Benefits/Benefits.js
@@ -1,5 +1,4 @@
 import { Fragment } from 'react';
-import Image from 'next/image'
 import {
   HeaderBenefitsWrapper,
   TextForBenefits,
