--- app/contexts/cartContext/cartContext.js
@@ -1,7 +1,0 @@
-"use client"
-
-import { createContext } from 'react'
-
-export const CartContext = createContext({})
-
-export default CartContext
