--- app/error.js
@@ -1,18 +1,0 @@
-'use client'
- 
-import { useEffect } from 'react'
- 
-export default function Error({ error, reset }) {
-  useEffect(() => {
-    console.error(error)
-  }, [error])
- 
-  return (
-    <div>
-      <h2>Something went wrong GLOBAL!</h2>
-      <button onClick={() => reset()}>
-        Try again
-      </button>
-    </div>
-  )
-}
