+++ app/error.tsx
@@ -1,18 +1,26 @@
 'use client'
 
+import React, {FC} from 'react'
 import { useEffect } from 'react'
+
+interface ErrorProps {
+  error: Error;
+  reset: () => void 
+}
  
-export default function Error({ error, reset }) {
+const Error: FC<ErrorProps> = ({ error, reset }) => {
   useEffect(() => {
     console.error(error)
   }, [error])
  
   return (
     <div>
-      <h2>Something went wrong COMPONENTS!</h2>
+      <h2>Something went wrong GLOBAL!</h2>
       <button onClick={() => reset()}>
         Try again
       </button>
     </div>
   )
 }
+
+export default Error;
