+++ app/layout.tsx
@@ -1,3 +1,4 @@
+import React, { ReactNode } from 'react';
 import localFont from "next/font/local";
 import HeaderComponent from './components/HeaderComponent/HeaderComponent'
 import FooterComponent from './components/FooterComponent/FooterComponent'
@@ -22,10 +23,10 @@ export const metadata = {
   robots: {
     index: true,
     follow: true,
-    googleBot: { // do we need it?
-      index: true,
-      follow: true,
-    },  
+    // googleBot: {  do we need it?
+    //   index: true,
+    //   follow: true,
+    // },  
   },
   link:{
     rel:"icon",
@@ -34,39 +35,44 @@ export const metadata = {
   }
 
 };
-  // robots.txt FILE ???
-  // Author and Copyright ?
-  // sitemap
+  //  robots.txt FILE ???
+  //  Author and Copyright ?
+  //  sitemap
+
+
+// interface RootLayoutProps {
+//   children: ReactNode;
+// }
+
   
-export default async function RootLayout({ children }) {
-  // const layoutData = await fetchProducts()
+export default async function RootLayout({ children }: { children: React.ReactNode }) {
+
+  //  const layoutData = await fetchProducts()
   let layoutData = [];
   try {
     layoutData = await getProduct(); 
-    // console.log('Fetched products in RootLayout:', layoutData);
-
   } catch (error) {
     console.error('Error in RootLayout fetching products:', error);
   }
   
-  console.log('ROOT context', layoutData?.[0]?.stock, layoutData?.[2]?.stock)
+   console.log('ROOT context', layoutData?.[0]?.stock, layoutData?.[2]?.stock)
 
   return (
     <html lang="es">
       <body className={`${mohave.variable}`}>
-        <StyledRegistry>
-          <GlobalStyle />
-            <CartProvider layoutData={layoutData}>
-              <HeaderComponent />
-                {children}
-              <FooterComponent />
-            </CartProvider>
-        </StyledRegistry>
+         <StyledRegistry>
+           <GlobalStyle />
+             <CartProvider layoutData={layoutData}>
+               <HeaderComponent />
+                 {children}
+               <FooterComponent />
+             </CartProvider>
+         </StyledRegistry>
       </body>
     </html>
   );
 }
 
-// REV 2
-// export const dynamic = 'force-dynamic';
-// export const revalidate = 5;
+//  REV 2
+//  export const dynamic = 'force-dynamic';
+//  export const revalidate = 5;
